+++ include/phase4/engine/board/operators/pawn_operator.h
@@ -7,12 +7,12 @@
 #include <phase4/engine/moves/moves_generator.h>
 
 #include <phase4/engine/common/bitset.h>
-#include <phase4/engine/common/evaluation_constants.h>
 #include <phase4/engine/common/piece_color.h>
 #include <phase4/engine/common/piece_type.h>
 #include <phase4/engine/common/position_constants.h>
 #include <phase4/engine/common/square.h>
 #include <phase4/engine/common/wall_operations.h>
+#include <phase4/engine/score/evaluation_constants.h>
 
 #include <cstdint>
