+++ include/phase4/engine/score/piece_square_tables/knight_tables.h
@@ -1,29 +1,29 @@
-#ifndef PHASE4_ENGINE_AI_SCORE_PIECE_SQUARE_TABLES_BISHOP_TABLES_H
-#define PHASE4_ENGINE_AI_SCORE_PIECE_SQUARE_TABLES_BISHOP_TABLES_H
+#ifndef PHASE4_ENGINE_SCORE_PIECE_SQUARE_TABLES_KNIGHT_TABLES_H
+#define PHASE4_ENGINE_SCORE_PIECE_SQUARE_TABLES_KNIGHT_TABLES_H
 
-#include <phase4/engine/ai/score/piece_square_tables/table_operations.h>
+#include <phase4/engine/score/piece_square_tables/table_operations.h>
 
 #include <phase4/engine/common/game_phase.h>
 
 #include <array>
 #include <cstdint>
 
-namespace phase4::engine::ai::score::piece_square_tables {
+namespace phase4::engine::score::piece_square_tables {
 
-class BishopTables {
+class KnightTables {
 public:
-	static constexpr int32_t O0 = 15; // Center
-	static constexpr int32_t O1 = 10; // Center
-	static constexpr int32_t O2 = 5; // Center
-	static constexpr int32_t O3 = 0; // Original fields
-	static constexpr int32_t O4 = -5; // Edge
-	static constexpr int32_t O5 = -15; // Corners
+	static const int32_t O0 = 15; // Center
+	static const int32_t O1 = 15; // Center
+	static const int32_t O2 = 15; // Center
+	static const int32_t O3 = 0; // Original fields
+	static const int32_t O4 = -10; // Edge
+	static const int32_t O5 = -15; // Corners
 
-	static constexpr int32_t E0 = 15; // Center
-	static constexpr int32_t E1 = 15; // Center
-	static constexpr int32_t E2 = 15; // Center
-	static constexpr int32_t E3 = -10; // Edge
-	static constexpr int32_t E4 = -15; // Corners
+	static const int32_t E0 = 10; // Center
+	static const int32_t E1 = 5; // Center
+	static const int32_t E2 = 0; // Center
+	static const int32_t E3 = -10; // Edge
+	static const int32_t E4 = -15; // Corners
 
 	static constexpr std::array<std::array<std::array<int32_t, 64>, 2>, 2> build() {
 		std::array<std::array<int32_t, 64>, 2> pattern = {
@@ -31,11 +31,11 @@ public:
 			{
 					{ O5, O4, O4, O4, O4, O4, O4, O5,
 							O4, O2, O2, O2, O2, O2, O2, O4,
-							O4, O2, O2, O1, O1, O2, O2, O4,
 							O4, O2, O1, O1, O1, O1, O2, O4,
 							O4, O1, O1, O0, O0, O1, O1, O4,
+							O4, O2, O0, O0, O0, O0, O2, O4,
 							O4, O1, O0, O0, O0, O0, O1, O4,
-							O4, O0, O2, O1, O1, O2, O0, O4,
+							O4, O1, O2, O2, O2, O2, O1, O4,
 							O5, O4, O3, O4, O4, O3, O4, O5 },
 					// Ending
 					{
@@ -54,7 +54,6 @@ public:
 			std::array<std::array<int32_t, 64>, 2>{
 					TableOperations::flipVertically(pattern[common::GamePhase::OPENING]),
 					TableOperations::flipVertically(pattern[common::GamePhase::ENDING]) },
-
 			// Black
 			std::array<std::array<int32_t, 64>, 2>{
 					TableOperations::flipHorizontally(pattern[common::GamePhase::OPENING]),
@@ -63,5 +62,5 @@ public:
 	}
 };
 
-} //namespace phase4::engine::ai::score::piece_square_tables
+} //namespace phase4::engine::score::piece_square_tables
 #endif
